refactor(podman): extract getBundledPodmanVersion interface to a file
