+++ extensions/podman/packages/extension/src/installer/base-installer.ts
@@ -18,7 +18,7 @@
 import type { InstallCheck } from '@podman-desktop/api';
 import { compare } from 'compare-versions';
 
-import { getBundledPodmanVersion } from '../utils/podman-install';
+import { getBundledPodmanVersion } from '../utils/podman-bundled';
 import type { Installer } from './installer';
 
 export abstract class BaseInstaller implements Installer {
